Rewrite: Add judgments, terms, contexts, and context containment
Context containment will hopefully allow us to forgo an explicit
notion of lifting a term/judgment into a context with a new
variable as part of the core. I say 'hopefully' because I suspect
it may make writing the equivalence rule for lambda application
hard/impossible.

A rewrite turned out to be important for two key reasons:

* Contexts being explicitly indexed by their length severely
  complicated typechecking, with no obvious benefit. If desired,
  proofs about the length transformation properties of functions
  on contexts can be written, but it's not obvious that any such
  functions (besides (::)) will ever be needed.
* Having terms for types and values live in the same type family
  will allow for a much simpler notion of transport over
  equivalences once we have those