Simplify term indices, delay universe lift inference
Instead of being indexed by their type, terms are now only indexed by
whether or not they are a type (and if so, of what universe level).

Instead of assigning any containing level that matches the relevant
constraints, assign the smallest level that does and simply make it
legal to use a term of a lower level where one of a higher level is
expected.